requirements.txt
discord.py==1.7.3 
discord-py-slash-command==3.0.0
pytz